4 - Jogo estável com proto salvamento

+++ ListaObservadores.cpp
@@ -85,29 +85,29 @@ ListaObservadores::~ListaObservadores()
 
 void ListaObservadores::notificarTeclaPressionada(const Keyboard::Key tecla)
 {
-	for (int i = 0; i < listaObservadores.size(); i++)
+	for (int i = 0; i < (int) listaObservadores.size(); i++)
 		if (listaObservadores[i]->getAtivado())
 			listaObservadores[i]->teclaPressionada(tecla);
 }
 
 void ListaObservadores::notificarTeclaSolta(const Keyboard::Key tecla)
 {
-	for (int i = 0; i < listaObservadores.size(); i++)
+	for (int i = 0; i < (int) listaObservadores.size(); i++)
 		if (listaObservadores[i]->getAtivado())
 			listaObservadores[i]->teclaSolta(tecla);
 }
 
 void ListaObservadores::notificarMouseSolta(Event::MouseMoveEvent mouse)// Mouse::Button botaoMouse, 
 {
-	for (int i = 0; i < listaObservadores.size(); i++)
+	for (int i = 0; i < (int) listaObservadores.size(); i++)
 		if (listaObservadores[i]->getAtivado())
-			listaObservadores[i]->botaoMouseSolta(Vector2f(mouse.x, mouse.y)); //botaoMouse,  Vector2f(mouse.x, mouse.y)
+			listaObservadores[i]->botaoMouseSolta(Vector2f((float) mouse.x, (float) mouse.y)); //botaoMouse,  Vector2f(mouse.x, mouse.y)
 }
 
 
 void ListaObservadores::desativarObservadores()
 {
-	for (int i = 0; i < listaObservadores.size(); i++)
+	for (int i = 0; i < (int) listaObservadores.size(); i++)
 		if (listaObservadores[i]->getAtivado())
 			listaObservadores[i]->setAtivado(false);
 }
@@ -124,7 +124,7 @@ void ListaObservadores::adicionarObservador(Observador* observador)
 
 void ListaObservadores::ativaObservador(IDs id)
 {
-	for (int i = 0; i < listaObservadores.size(); i++)
+	for (int i = 0; i < (int) listaObservadores.size(); i++)
 	{
 		if (listaObservadores[i]->getID() == id)
 		{
